requirements.txt
# Core Framework
dynaconf>=3.2.6
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# Google ADK & LLM
google-adk>=1.26.0
litellm>=1.53.0

# MCP (Model Context Protocol)
fastmcp>=3.0.0
mcp>=1.2.0
httpx-sse>=0.4.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.10
redis>=5.2.0

# Mathematics & Scientific Computing
numpy>=2.2.0
scipy>=1.14.0
networkx>=3.4.0
scikit-learn>=1.6.0
pulp>=2.9.0

# Semantic RAG — ChromaDB vector store + sentence-transformers embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data Processing
pandas>=2.2.0

# HTTP & Async
httpx>=0.28.0
aiohttp>=3.11.0
tenacity>=9.0.0

# Logging
loguru>=0.7.3

# UI
streamlit>=1.41.0
plotly>=5.24.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
hypothesis>=6.112.0    # property-based testing

# Dev Tools
python-dotenv>=1.0.1
click>=8.1.0
rich>=13.9.0
